+++ src/main/java/frc/robot/commands/Drive/DriveSimple.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.Drive;
 
 import com.frcteam3255.joystick.SN_F310Gamepad;
